+++ src/react_native/index.ts
@@ -4,8 +4,8 @@ import CborReader from 'cbor-js';
 import { Buffer } from 'buffer';
 
 import { ExtendedConfiguration, PlatformConfiguration } from '../core/interfaces/configuration';
-import { WebReactNativeTransport } from '../transport/web-react-native-transport';
 import { stringifyBufferKeys } from '../core/components/stringify_buffer_keys';
+import { ReactNativeTransport } from '../transport/react-native-transport';
 import { makeConfiguration } from '../core/components/configuration';
 import { PubNubFileParameters } from '../file/modules/react-native';
 import { TokenManager } from '../core/components/token_manager';
@@ -72,7 +72,7 @@ export default class PubNub extends PubNubCore<null, PubNubFileParameters> {
     const transportMiddleware = new PubNubMiddleware({
       clientConfiguration,
       tokenManager,
-      transport: new WebReactNativeTransport(fetch, clientConfiguration.keepAlive, clientConfiguration.logVerbosity!),
+      transport: new ReactNativeTransport(clientConfiguration.keepAlive, clientConfiguration.logVerbosity!),
     });
 
     super({